升级安装脚本。

+++ uninstall.bat
@@ -3,39 +3,51 @@
 
 set mmpath=C:\Windows
 
+set sourcefile=mm.bat
 set mmfile=%mmpath%\mm.bat
 call :uninstall
 
+set sourcefile=pymm.bat
 set mmfile=%mmpath%\pymm.bat
 call :uninstall
 
+set sourcefile=pymake.bat
 set mmfile=%mmpath%\pymake.bat
 call :uninstall
 
+set sourcefile=pyenv.bat
 set mmfile=%mmpath%\pyenv.bat
 call :uninstall
 
+set sourcefile=pyvc.bat
 set mmfile=%mmpath%\pyvc.bat
 call :uninstall
 
+set sourcefile=pycmd.bat
 set mmfile=%mmpath%\pycmd.bat
 call :uninstall
 
+set sourcefile=pyexecvp.bat
 set mmfile=%mmpath%\pyexecvp.bat
 call :uninstall
 
+set sourcefile=pyccvp.bat
 set mmfile=%mmpath%\pyccvp.bat
 call :uninstall
 
+set sourcefile=pyinfo.bat
 set mmfile=%mmpath%\pyinfo.bat
 call :uninstall
 
+set sourcefile=pypaths.bat
 set mmfile=%mmpath%\pypaths.bat
 call :uninstall
 
+echo –∂‘ÿ≥…π¶°£
 exit /b 0
 
 ::================================================================
 :uninstall
-call del /q /f %mmfile%
+if exist "%mmfile%" call del /q /f %mmfile%
+echo –∂‘ÿ %sourcefile%
 goto :eof
